requirements.txt
mkdocs>=1.4.0
mkdocs-material>=9.0.0
pyyaml>=6.0
